+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Added raises attribute to the docstrings.
 * Added OpenAPI docs.
 * Refactored all SCSS classes to BEM naming.